+++ src/model/discord.rs
@@ -3,7 +3,7 @@ use serde::{Deserialize, Serialize};
 #[derive(Serialize, Deserialize, Clone)]
 pub struct DiscordGuildDto {
     pub id: i32,
-    pub guild_id: i32,
+    pub guild_id: i64,
     pub name: String,
     pub icon_hash: Option<String>,
 }
